Refactor features
